move Footer component out of Geral subfolder
