+++ website/src/Components/Admin/ConfigManager.js
@@ -31,7 +31,7 @@ import ConfirmLeave from 'Components/Notifications/Geral/ConfirmLeave';
 import Notification from 'Components/Notifications/Geral/Notification';
 import ConfirmActionPopup from 'Components/Form/Geral/ConfirmActionPopup';
 import CheckboxList from 'Components/Form/Geral/CheckboxList';
-import Footer from 'Components/Footer/Geral/Footer';
+import Footer from 'Components/Footer/Footer';
 
 const API_URL = `${window.location.protocol}//${window.location.host}/${process.env.REACT_APP_API_URL}`;
 const ADMIN_HOME = (process.env.REACT_APP_BASENAME !== null && process.env.REACT_APP_BASENAME !== "")
