+++ website/src/Components/Admin/StorageManager.js
@@ -20,7 +20,7 @@ import Notification from 'Components/Notifications/Geral/Notification';
 import ConfirmActionPopup from 'Components/Form/Geral/ConfirmActionPopup';
 import CheckboxList from 'Components/Form/Geral/CheckboxList';
 import TooltipIcon from 'Components/TooltipIcon/Geral/TooltipIcon';
-import Footer from 'Components/Footer/Geral/Footer';
+import Footer from 'Components/Footer/Footer';
 
 const API_URL = `${window.location.protocol}//${window.location.host}/${process.env.REACT_APP_API_URL}`;
 const ADMIN_HOME = (process.env.REACT_APP_BASENAME !== null && process.env.REACT_APP_BASENAME !== "")
